+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.4
+* All of the package dependencies are supported in the latest version
+* Adds `ChangeNotifier` class to include parameters to change theme using through the 
+  implementation of `ChangeThemeSwitchWidget` & `ChangeThemeChoiceListTileWidget`.
+* Adds `ChangeThemeChoiceListTileWidget` Widget to switching between `Light` `Dark` & `System` Theme.
+* Adds `ChangeThemeSwitchWidget` Widget to switching between `Light` & `Dark` Theme.
+* Fixes INFO: `useInheritedMediaQuery` is deprecated and shouldn't be used. Remove this parameter as it is now ignored. MaterialApp never introduces its own `MediaQuery;` the View widget takes care of that. This feature was deprecated after v3.7.0-29.0.pre.
+
+
 ## 0.0.3
 * All of the package dependencies are supported in the latest version.